refactor(ui-react): extract web and diff previews
